Final backend commit

+++ src/main/java/com/logincontroller/filmrentalsystem/dto/CategoryDTO.java
@@ -1,12 +1,13 @@
 package com.logincontroller.filmrentalsystem.dto;
 
 import lombok.Data;
+
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data
 public class CategoryDTO {
-    private Integer categoryId;
+    private Byte categoryId;
     private String name;
     private LocalDateTime lastUpdate;
     private List<String> filmTitles;
